\name{NetworkReg}
\alias{NetworkReg}
\docType{package}
\title{
  Regression Model on Network-Linked Data with Statistical Inference
}
\description{
Linear regression model with nonparametric network effects on network-linked observations. The model is proposed by Le and Li (2020) <arXiv:2007.00803> on observations that are connected by a network or similar relational data structure. The model does not assume that the relational data or network structure to be precisely observed; thus, the method is provably robust to a certain level of perturbation of the network structure. The package contains the estimation and inference function for the model.
}
\details{
\tabular{ll}{
Package: \tab NetworkReg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2020-07-02\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Can M. Le and  Tianxi Li.\cr

Maintainer: Tianxi Li <tianxili@virginia.edu>
}
\references{
Can M. Le and Tianxi Li.  Linear regression and its inference on noisy network-linked data. arXiv:2007.00803
}
\keyword{ package }
