\name{NetworkReg-package}
\alias{NetworkReg-package}
\docType{package}
\title{
  Generalized Linear Regression Models on Network-Linked Data with Statistical Inference
}
\description{
Linear regression model with nonparametric network effects on network-linked observations. The model is proposed by Le and Li (2022) <arXiv:2007.00803> on observations that are connected by a network or similar relational data structure. The model does not assume that the relational data or network structure to be precisely observed; thus, the method is provably robust to a certain level of perturbation of the network structure. The package contains the estimation and inference function for the model.
}
\details{
\tabular{ll}{
Package: \tab NetworkReg\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2024-10-10\cr
License: \tab GPL (>= 2)\cr
}

}
\references{
Le, C. M., & Li, T. (2022). Linear regression and its inference on noisy network-linked data. Journal of the Royal Statistical Society Series B: Statistical Methodology, 84(5), 1851-1885.

Wang J, Le C M, Li T. Perturbation-Robust Predictive Modeling of Social Effects by Network Subspace Generalized Linear Models. arXiv preprint arXiv:2410.01163, 2024.
}
\author{
Jianxiang Wang, Can M. Le, and Tianxi Li.\cr
Maintainer: Jianxiang Wang <jw1881@scarletmail.rutgers.edu>
}
\keyword{ package }


