% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP_Inf.R
\name{net.gen.from.P}
\alias{net.gen.from.P}
\title{generates a network from the given connection probability}
\usage{
net.gen.from.P(P, mode = "undirected")
}
\arguments{
\item{P}{connection probability between nodes}

\item{mode}{"undirected" (default) if the network is undirected, so the
adjacency matrix will be symmetric with only upper diagonal entries
being generated as independent Bernoulli. Otherwise, the adjacency matrix gives independent Bernoulli everywhere.}
}
\value{
An adjacency matrix
}
\description{
Generates an adjacency matrix from a given probability matrix,
according independent Bernoulli -- the so-called inhomogeneous
Erdos-Renyi model. It is used to generate new networks from a given model.
}
