\name{NetworkReg}
\alias{NetworkReg}
\docType{package}
\title{
  Regression Model on Network-Linked Data with Statistical Inference
}
\description{
Linear regression model with nonparametric network effects on network-linked observations. The model is proposed by Le and Li (2022) <arXiv:2007.00803> on observations that are connected by a network or similar relational data structure. The model does not assume that the relational data or network structure to be precisely observed; thus, the method is provably robust to a certain level of perturbation of the network structure. The package contains the estimation and inference function for the model.
}
\details{
\tabular{ll}{
Package: \tab NetworkReg\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2024-02-02\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Can M. Le and  Tianxi Li.\cr

Maintainer: Tianxi Li <tianxili@umn.edu>
}
\references{
Can M. Le and Tianxi Li.  Le, C. M., & Li, T. (2022). Linear regression and its inference on noisy network-linked data. Journal of the Royal Statistical Society Series B: Statistical Methodology, 84(5), 1851-1885.
}
\keyword{ package }
